pyproject.toml
# Описание работы с poetry в ./docs/poetry.md

[tool.poetry]
name = "lizaalert"
version = "0.1.0"
description = ""
authors = ["Lizaalert-backend-team"]

[tool.poetry.dependencies]
python = "^3.7.12"
Django = "3.2.12"
djangorestframework = "3.13.1"
environs = "9.5.0"
gunicorn = "20.1.0"
pydantic = {version = "1.9.0", extras=["dotenv"]}
psycopg2-binary = "2.9.3"
easy-thumbnails = "2.8.1"
phonenumbers = "8.12.45"
django-phonenumber-field = "6.1.0"
drf-yasg = "1.20.0"
dj-rest-auth = "2.2.4"
django-allauth = "0.50.0"
djangorestframework-simplejwt = "5.1.0"

[tool.poetry.dev-dependencies]
flake8 = "4.0.1"
isort = "5.10.1"
black = "22.3.0"
pre-commit = "2.18.1"
pytest = "7.1.2"
pytest-django = "^4.5.2"
requests = "2.27.1"
pytest-cov = "3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ["py310"]
extend-exclude = '\/migrations\/'